Continuing to get the lexer to work as a whole

+++ syntax/src/token.rs
@@ -10,27 +10,27 @@ pub enum WhitespaceType {
 pub enum Token<'a> {
     Start,
     End,
-    Bang(usize, u32, u32),
-    Dollar(usize, u32, u32),
-    Amp(usize, u32, u32),
-    OpenParen(usize, u32, u32),
-    CloseParen(usize, u32, u32),
-    Spread(usize, u32, u32),
-    Colon(usize, u32, u32),
-    Equals(usize, u32, u32),
-    At(usize, u32, u32),
-    OpenSquare(usize, u32, u32),
-    CloseSquare(usize, u32, u32),
-    OpenBrace(usize, u32, u32),
-    CloseBrace(usize, u32, u32),
-    Pipe(usize, u32, u32),
-    Name(usize, u32, u32, &'a str),
-    Int(usize, u32, u32, i64),
-    Float(usize, u32, u32, f64),
-    Str(usize, u32, u32, &'a str),
-    BlockStr(usize, u32, u32, &'a str),
-    Comment(usize, u32, u32, &'a str),
-    Whitespace(usize, u32, u32, WhitespaceType),
+    Bang(usize, usize, usize),
+    Dollar(usize, usize, usize),
+    Amp(usize, usize, usize),
+    OpenParen(usize, usize, usize),
+    CloseParen(usize, usize, usize),
+    Spread(usize, usize, usize),
+    Colon(usize, usize, usize),
+    Equals(usize, usize, usize),
+    At(usize, usize, usize),
+    OpenSquare(usize, usize, usize),
+    CloseSquare(usize, usize, usize),
+    OpenBrace(usize, usize, usize),
+    CloseBrace(usize, usize, usize),
+    Pipe(usize, usize, usize),
+    Name(usize, usize, usize, &'a str),
+    Int(usize, usize, usize, i64),
+    Float(usize, usize, usize, f64),
+    Str(usize, usize, usize, &'a str),
+    BlockStr(usize, usize, usize, &'a str),
+    Comment(usize, usize, usize, &'a str),
+    Whitespace(usize, usize, usize, WhitespaceType),
 }
 
 impl<'a> Token<'a> {
@@ -43,9 +43,6 @@ use std::fmt;
 
 impl<'a> fmt::Display for Token<'a> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // match self {
-        //     Start | End =>
-        // }
         write!(f, "Token<{:?}>", self)
     }
 }
